correcting storyboard warnings fix bug with the observation custom time filter not setting if you do not change the time unit

+++ Mage/LocationFilterTableViewController.m
@@ -42,7 +42,7 @@
         UITextField *timeNumberField = (UITextField *) [cell viewWithTag:200];
         
         timeUnitControl.selectedSegmentIndex = self.customTimeUnit;
-        timeNumberField.text = [NSString stringWithFormat:@"%ld", self.customTimeNumber];
+        timeNumberField.text = [NSString stringWithFormat:@"%ld", (long)self.customTimeNumber];
     }
     
     return cell;
